feat : 알림 요소 컴포넌트화
